Refactors components into folders
