+++ front-end/App.js
@@ -2,7 +2,7 @@ import { StatusBar } from "expo-status-bar";
 import React, { useContext, useEffect, useState } from "react";
 import { useFonts } from "expo-font";
 
-import Header from "./Components/Header";
+import Header from "./Components/Header/Header";
 import { UserContext } from "./Contexts/UserContext";
 import PublicNavigator from "./Components/Navigators/PublicNavigator";
 import PlayerNavigator from "./Components/Navigators/PlayerNavigator";
