+++ front-end/Components/Home/LeagueTable.js
@@ -1,4 +1,3 @@
-import { useEffect } from "react";
 import { StyleSheet } from "react-native";
 import { DataTable } from "react-native-paper";
