+++ front-end/Components/Utils/emailValidator.js
