+++ front-end/Components/Utils/nameValidator.js
