+++ front-end/Components/Utils/passwordValidator.js
